batch/requirements.txt
bar-chart-race==0.1.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
japanize-matplotlib==1.1.3
kiwisolver==1.4.6
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==74.1.1
six==1.16.0
SQLAlchemy==2.0.33
typing_extensions==4.12.2
tzdata==2024.1
